+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This release adds the following new features:
 - Drag and drop of files and conversations into the tab column view.
 - Added new icons for the file view and made them scalable if you zoom in and out.
 - Added the ability to delete files in the file view by right clicking and selecting "Delete File".
+- Added a Kotlin lexer/parser for syntax highlighting.
 
 Te release fixes the following bugs: